feat: move profile settings to Settings
