+++ apps/ui/src/views/Create.vue
@@ -202,15 +202,24 @@ watchEffect(() => setTitle('Create space'));
       </div>
       <div class="flex-1">
         <div class="mt-8 lg:mt-0">
-          <FormProfile
-            v-if="currentPage === 'profile'"
-            :form="metadataForm"
-            :treasuries-value="metadataForm.treasuries"
-            :delegations-value="metadataForm.delegations"
-            @treasuries="v => (metadataForm.treasuries = v)"
-            @delegations="v => (metadataForm.delegations = v)"
-            @errors="v => handleErrors('profile', v)"
-          />
+          <template v-if="currentPage === 'profile'">
+            <h3 class="mb-4">Space profile</h3>
+            <FormSpaceProfile
+              :form="metadataForm"
+              @errors="v => handleErrors('profile', v)"
+            />
+            <div class="s-box p-4 -mt-6">
+              <FormSpaceTreasuries
+                :treasuries-value="metadataForm.treasuries"
+                @treasuries="v => (metadataForm.treasuries = v)"
+              />
+              <FormSpaceDelegations
+                class="mt-2"
+                :delegations-value="metadataForm.delegations"
+                @delegations="v => (metadataForm.delegations = v)"
+              />
+            </div>
+          </template>
           <FormNetwork
             v-else-if="currentPage === 'network'"
             v-model="selectedNetworkId"
